Move rootReducer to root of state
Allows us to start defining the state in a more modular way, i.e.
actions, reducer and selectors of one part of the state in one
file/place.
Refs #572.
