+++ app/state/rootReducer.js
@@ -1,3 +1,19 @@
-import rootReducer from './reducers';
+import { combineReducers } from 'redux';
+import { reducer as formReducer } from 'redux-form';
+import { intlReducer } from 'react-intl-redux';
 
-export default rootReducer;
+import apiReducers from './reducers/api';
+import authReducer from './reducers/authReducer';
+import dataReducer from './reducers/dataReducer';
+import notificationsReducer from './reducers/notificationsReducer';
+import uiReducers from './reducers/ui';
+
+export default combineReducers({
+  api: apiReducers,
+  auth: authReducer,
+  data: dataReducer,
+  form: formReducer,
+  intl: intlReducer,
+  notifications: notificationsReducer,
+  ui: uiReducers,
+});
