+++ app/utils/testUtils.js
@@ -10,7 +10,7 @@ import React from 'react';
 import { IntlProvider } from 'react-intl';
 import { CALL_API } from 'redux-api-middleware';
 
-import rootReducer from 'state/reducers';
+import rootReducer from 'state/rootReducer';
 import enMessages from 'i18n/messages/en.json';
 
 const testMessages = mapValues(enMessages, (value, key) => key);
